feat: fixes audit L1
Reimpmlement SafeCast for GLRDT. This was initially removed as each use
would have been protected by the maximum supply, or unfeasible to hit in
realistic scenarios (such as a length of an array exceeding a uint32).

To be on the safe side these have been re-implemented as it will also
help those reading the contract to have confidence that the GLRDT
contract has not been meaningfully changed from OZ's ERC20Votes.sol.

+++ src/interfaces/IGovernanceLockedRevenueDistributionToken.sol
@@ -10,8 +10,8 @@ interface IGovernanceLockedRevenueDistributionToken {
      */
     struct Checkpoint {
         uint32 fromBlock;
-        uint112 votes;
-        uint112 shares;
+        uint96 votes;
+        uint96 shares;
     }
 
     /*░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
@@ -56,7 +56,7 @@ interface IGovernanceLockedRevenueDistributionToken {
     function userCheckpoints(address account_, uint256 pos_)
         external
         view
-        returns (uint32 fromBlock, uint112 votes, uint112 shares);
+        returns (uint32 fromBlock, uint96 votes, uint96 shares);
 
     /*░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
     ░░░░                         Public Functions                          ░░░░
